+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 #### Added
 - added C-Api which can be used to extend RapidFuzz from different Python modules using any
   programming language which allows the usage of C-Apis (C/C++/Rust)
+- added new scorers in `rapidfuzz.algorithm.*`
+  - port existing ports to this new api
+  - add Indel distance
 
 #### Changed
 - when the result of `string_metric.levenshtein` or `string_metric.hamming` is below max